Efficiently pack data
Moves time fields to uint64, except for auction length which is uint32 
(signifying a max auction length of ~68 years).

Tweaks proposal type to uint16 and changes the commonProposalBit from 
most significant to least significant in the more significant byte. If 
used as a bit flag, there's 8 possible enum types which should be more 
than plenty (right now we have 2 and I don't predict we'll need a 
third). If used as an independent uint8, there's 255 possible types. By 
using the least significant bit this time, we can decide in the future 
if it's a bit flag or an int.

+++ contracts/erc20/FrabricERC20.sol
@@ -48,8 +48,10 @@ contract FrabricERC20 is OwnableUpgradeable, PausableUpgradeable, DividendERC20,
     _setWhitelisted(msg.sender, keccak256("Initializer"));
 
     // Make sure the supply is within bounds
-    // The DAO code sets an upper bound of int256.max
-    // Uniswap and more frequently use uint112 which is still a very reasonable bound
+    // The DAO code sets an upper bound of signed<int>.max
+    // Uniswap and more frequently use uint112 which is a perfectly functional bound
+    // The DAO code accordingly uses int128 which maintains storage efficiency
+    // while supporting the full uint112 range
     if (supply > uint256(type(uint112).max)) {
       revert SupplyExceedsUInt112(supply);
     }
@@ -126,7 +128,13 @@ contract FrabricERC20 is OwnableUpgradeable, PausableUpgradeable, DividendERC20,
         _transfer(person, auction, amount);
 
         // List the transferred tokens
-        IAuction(auction).listTransferred(address(this), dexToken, person, block.timestamp + (i * (1 weeks)), 1 weeks);
+        IAuction(auction).listTransferred(
+          address(this),
+          dexToken,
+          person,
+          uint64(block.timestamp) + (i * (1 weeks)),
+          1 weeks
+        );
       }
       _removal = false;
     }
